+++ CHANGELOG.md
@@ -9,6 +9,7 @@ ENHANCEMENTS:
 BUG FIXES:
 
 * Refresh Plan not empty for `sonatypeiq_source_control` due to Token masking [GH-50]
+* Data Source `sonatypeiq_config_saml` no longer errors when there is no SAML configuration present [GH-53]
 
 
 ## 0.10.0 July 14, 2025